💚 Fix cookbook build
- Move template to file and use noop() from rxjs
- Move example component and use noop from rxjs
