+++ apps/cookbook/src/app/examples/card-example/card-example.module.ts
@@ -4,7 +4,7 @@ import { NgModule } from '@angular/core';
 import { KirbyModule } from '@kirbydesign/designsystem';
 
 import { CardExampleComponent } from './card-example.component';
-import { CardBackgroundImageExampleComponent } from './examples/card-background-image-example.component';
+import { CardBackgroundImageExampleComponent } from './examples/card-background-image-example/card-background-image-example.component';
 import { CardClickableExampleComponent } from './examples/card-clickable-example/card-clickable-example.component';
 import { CardCssBackgroundImageExampleComponent } from './examples/card-css-background-image-example/card-css-background-image-example.component';
 import { CardElevationsExampleComponent } from './examples/card-elevations-example/card-elevations-example.component';
