+++ apps/cookbook/src/app/examples/card-example/examples/card-background-image-example/card-background-image-example.component.html
@@ -1,0 +1,13 @@
+<kirby-card
+  [hasPadding]="true"
+  backgroundImageUrl="https://images.unsplash.com/photo-1560840067-ddcaeb7831d2"
+  themeColor="dark"
+  (click)="noop()"
+>
+  <h3>Example using input property to set background</h3>
+  <p>
+    Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae
+    necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo accusamus cumque, in
+    quia itaque cupiditate ratione repellat!
+  </p>
+</kirby-card>
