+++ apps/cookbook/src/app/examples/card-example/examples/card-background-image-example/card-background-image-example.component.ts
@@ -1,0 +1,20 @@
+import { Component } from '@angular/core';
+import { noop } from 'rxjs';
+
+const config = {
+  selector: 'cookbook-card-background-image-example',
+  style: `kirby-card {
+  min-height: 300px;
+}
+`,
+};
+
+@Component({
+  selector: config.selector,
+  templateUrl: './card-background-image-example.component.html',
+  styles: [config.style],
+})
+export class CardBackgroundImageExampleComponent {
+  style: string = config.style;
+  noop: () => void = noop;
+}
