+++ apps/cookbook/src/app/examples/card-example/examples/card-clickable-example/card-clickable-example.component.html
@@ -1,4 +1,4 @@
-<kirby-card (click)="function()">
+<kirby-card (click)="noop()">
   <kirby-card-header title="Clickable"></kirby-card-header>
   <div class="card-content">
     <p>
