+++ apps/cookbook/src/app/examples/card-example/examples/card-css-background-image-example/card-css-background-image-example.component.html
@@ -1,0 +1,8 @@
+<kirby-card [hasPadding]="true" themeColor="dark" (click)="noop()">
+  <h3>Example using custom css property to set background</h3>
+  <p>
+    Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae
+    necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo accusamus cumque, in
+    quia itaque cupiditate ratione repellat!
+  </p>
+</kirby-card>
