+++ apps/cookbook/src/app/examples/card-example/examples/card-css-background-image-example/card-css-background-image-example.component.ts
@@ -1,17 +1,8 @@
 import { Component } from '@angular/core';
+import { noop } from 'rxjs';
 
 const config = {
   selector: 'cookbook-card-css-background-image-example',
-  template: `
-<kirby-card [hasPadding]="true" themeColor="dark" (click)="function()">
-  <h3>
-    Example using custom css property to set background
-  </h3>
-  <p> 
-    Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo accusamus cumque, in quia itaque cupiditate ratione repellat!
-  </p>
-</kirby-card>
-  `,
   style: `@use '~@kirbydesign/core/src/scss/utils';
 
 kirby-card {
@@ -37,10 +28,10 @@ kirby-card {
 
 @Component({
   selector: config.selector,
-  template: config.template,
+  templateUrl: './card-css-background-image-example.component.html',
   styleUrls: ['./card-css-background-image-example.component.scss'],
 })
 export class CardCssBackgroundImageExampleComponent {
-  template: string = config.template;
   style: string = config.style;
+  noop: () => void = noop;
 }
